requirements.txt
mypy==0.790
parameterized==0.7.4
python-vlc==3.0.11115
websockets==8.1
absl-py==0.11.0
pydantic==1.7.3